remove kometa's overlay label when uploading posters
kometa relies on an `Overlay` label to tell it what it has
previously processed.  If a poster changes outside of kometa
itself it won't detect this and the previously applied overlays
will be gone and never updated.  By removing the label, Kometa
will detect this on its next run and reapply overlays as needed.